Updates for GCC 4.3
git-svn-id: https://svn.zsnes.com/rep/zsnes/trunk@5233 8cdf4848-9e13-0410-98e6-c9478ff1fb58

+++ src/parsegen.cpp
@@ -1822,7 +1822,7 @@ void parser_generate(istream& psr_stream, ostream& c_stream, ostream& cheader_st
   }
 }
 
-int main(size_t argc, const char *const *const argv)
+int main(int argc, const char *const *const argv)
 {
   const char *cheader_file = 0;
   bool compile = false;
